+++ src/App.css
@@ -36,3 +36,7 @@
     transform: rotate(360deg);
   }
 }
+
+a { 
+    margin-right: 20px;
+ }
