+++ src/components/pages/about-page/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+
+export default function AboutPage() {
+    return <h1> AboutPage </h1>
+}
