+++ src/components/pages/configuration-page/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+
+export default function ConfigurationPage() {
+    return <h1> ConfigurationPage </h1>
+}
